pyproject.toml
[project]
name = "coding-interview"
version = "0.1.0"
description = "Category management API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0",
    "djangorestframework>=3.15",
    "psycopg2-binary>=2.9",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8",
    "mypy>=1.13",
    "django-stubs>=5.1",
    "djangorestframework-stubs>=3.15",
    "coverage>=7.6",
    "pre-commit>=4.0",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "DJ",  # flake8-django
    "PIE", # flake8-pie
    "PT",  # flake8-pytest-style
    "SIM", # flake8-simplify
]
ignore = [
    "E501", # line too long (handled by formatter)
    "DJ001", # null=True on CharField/TextField
]

[tool.ruff.lint.per-file-ignores]
"*/migrations/*.py" = ["E501", "N806", "N999"]
"*/tests/*.py" = ["S101", "S105", "S106", "PT009"]

[tool.mypy]
python_version = "3.11"
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings"

[tool.coverage.run]
source = ["api"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "config"]
